refactor(mcp): split edit tools into mcp/edit, move compute_diff to mcp/preview
