+++ dragonglass/mcp/__init__.py
@@ -1,14 +1,8 @@
 from __future__ import annotations
 
 from dragonglass._mod_replace import replace_modname
-from dragonglass.mcp.search import (
-    create_search_server,
-    delete_frontmatter_key_lines,
-    rebuild_note_with_frontmatter,
-    remove_inline_tags,
-    set_frontmatter_key_lines,
-    split_frontmatter_block,
-)
+from dragonglass.mcp.preview import compute_diff
+from dragonglass.mcp.search import create_search_server
 from dragonglass.mcp.telemetry import (
     MCPToolTelemetryEvent,
     ToolPhase,
@@ -17,6 +11,7 @@ from dragonglass.mcp.telemetry import (
 )
 
 for _sym in (  # noqa: RUF067
+    compute_diff,
     create_search_server,
     MCPToolTelemetryEvent,
     ToolPhase,
@@ -28,12 +23,8 @@ for _sym in (  # noqa: RUF067
 __all__ = [
     "MCPToolTelemetryEvent",
     "ToolPhase",
+    "compute_diff",
     "create_search_server",
-    "delete_frontmatter_key_lines",
     "drain_tool_events",
     "emit_tool_event",
-    "rebuild_note_with_frontmatter",
-    "remove_inline_tags",
-    "set_frontmatter_key_lines",
-    "split_frontmatter_block",
 ]
